refactor(front): rename feature folders
