+++ front/src/features/tables/EmptyLogs.tsx
