+++ front/src/features/tables/LogDetailsTable.tsx
